Update blog pages so images render
Promoted blog-search to replace blog
Updated to hide search bar for now, so it's just like the old blog pages, but with better structure and working images.
