+++ src/app/pages/blog/article-summary/article-summary.html
@@ -2,10 +2,12 @@
     <a class="link-image" href="/blog/{model.articleSlug}" aria-label="{model.altText}"
         style="{backgroundImage: `url(${model.image})`}"></a>
     <div class="text-block">
-        <h2 class="article-headline">{model.headline}</h2>
+        <h2 class="article-headline">
+            <a href="/blog/{model.articleSlug}">{model.headline}</a>
+        </h2>
         <h3 if="model.subheading">{model.subheading}</h3>
         <div class="article-blurb" data-html="body"></div>
-        <a class="go-to" href="/blog-with-search/{model.articleSlug}">read more</a>
+        <a class="go-to" href="/blog/{model.articleSlug}">read more</a>
         <div class="meta">
             <span class="author">{model.author}</span>
             <span class="date">{model.date}</span>
