--- src/app/pages/blog/article/formatted-as/feature.html
@@ -1,22 +1,0 @@
-<template args="model">
-    <div class="hero" style="{backgroundImage: `url(${model.coverUrl})`}"
-        role="img" aria-label="{model.coverAltText}"
-    >
-        <img class="strips" src="/images/components/strips.svg" height="10" alt="" role="presentation">
-    </div>
-    <article>
-        <h1>{model.title}</h1>
-        <h2 if="model.subheading">{model.subheading}</h2>
-        <div class="meta">
-            <p>
-                <span class="author">{model.author}</span>
-                <span class="date">{model.date}</span>
-            </p>
-        </div>
-        <div class="body"></div>
-        <div class="tags" if="model.tags.length">
-            <div class="tag" each="tag in model.tags">{tag}</div>
-        </div>
-        <div id="disqus_thread" class="disqus"></div>
-    </article>
-</template>
