--- src/app/pages/blog/article/formatted-as/synopsis.html
@@ -1,22 +1,0 @@
-<template args="model">
-    <a class="img" href="/blog/{model.articleSlug}"
-        style="{backgroundImage: `url(${model.coverUrl})`}"
-        role="img" aria-label="{model.coverAltText}"
-    ></a>
-    <div class="content">
-        <article class="excerpt">
-            <h2><a href="/blog/{model.articleSlug}">{model.title}</a></h2>
-            <h3 if="model.subheading && model.pinned">{model.subheading}</h3>
-            <div class="viewport">
-                <div class="body"></div>
-                <a class="go-to" href="/blog/{model.articleSlug}">read more</a>
-            </div>
-        </article>
-        <div class="meta">
-            <p>
-                <span class="author">{model.author}</span>
-                <span class="date">{model.date}</span>
-            </p>
-        </div>
-    </div>
-</template>
