--- src/app/pages/blog/blog.html
@@ -1,14 +1,0 @@
-<template args="model">
-    <div class="article page {model.articleSlug ? '' : 'hidden'}"></div>
-    <div class="boxed">
-        <div class="pinned"></div>
-    </div>
-    <div class="articles">
-        <div class="update-box" role="region">
-            <h3>Get updates in your inbox.</h3>
-            <a class="btn cta" href="http://www2.openstax.org/l/218812/2016-10-04/lvk">Sign Up</a>
-            {' or RSS '}<a class="rss-link" href="{model.rssUrl}"><i class="fa fa-rss-square"></i></a>
-        </div>
-        <div class="boxed container" role="complementary"></div>
-    </div>
-</template>
