+++ src/app/pages/blog/more-stories/more-stories.html
@@ -1,0 +1,2 @@
+<div class="cards boxed">
+</div>
