+++ src/app/pages/blog/more-stories/more-stories.js
@@ -2,6 +2,7 @@ import componentType from '~/helpers/controller/init-mixin';
 import ArticleSummary from '../article-summary/article-summary';
 import css from './more-stories.css';
 import {fetchFromCMS} from '~/helpers/controller/cms-mixin';
+import {description as template} from './more-stories.html';
 
 const cardSpec = {
     view: {
@@ -19,9 +20,13 @@ class Card extends ArticleSummary {
 }
 
 const spec = {
+    template,
     css,
     view: {
-        classes: ['more-stories', 'boxed']
+        classes: ['more-stories']
+    },
+    regions: {
+        cards: '.cards'
     }
 };
 
@@ -30,7 +35,7 @@ export default class extends componentType(spec) {
     onLoaded() {
         this.articles.forEach((model) => {
             delete model.subheading;
-            this.regions.self.append(new Card({model}));
+            this.regions.cards.append(new Card({model}));
         });
     }
