+++ src/app/pages/blog/more-stories/more-stories.scss
@@ -4,9 +4,15 @@
 $image-height: 24rem;
 
 .blog.page .more-stories {
-    display: grid;
-    grid-gap: 3rem;
-    grid-template-columns: repeat(auto-fit, minmax(36rem, 1fr));
+    background-color: ui-color(default);
+    padding-top: 2rem;
+    width: 100%;
+
+    .cards {
+        display: grid;
+        grid-gap: 3rem;
+        grid-template-columns: repeat(auto-fit, minmax(36rem, 1fr));
+    }
 
     .card {
         @extend %card;
