+++ src/app/pages/blog/pinned-article/pinned-article.js
