+++ src/app/pages/blog/search-bar/search-bar.js
