+++ src/app/pages/blog/search-results/search-results.scss
