+++ src/app/pages/blog/update-box/update-box.html
@@ -1,0 +1,6 @@
+<template args="model">
+    <h3>Get updates in your inbox.</h3>
+    <a class="btn cta" href="http://www2.openstax.org/l/218812/2016-10-04/lvk">Sign Up</a>
+    {' or RSS '}
+    <a class="rss-link" href="{model.rssUrl}"><i class="fa fa-rss-square"></i></a>
+</template>
