+++ src/app/pages/blog/update-box/update-box.scss
@@ -1,0 +1,37 @@
+@import 'pattern-library/core/pattern-library/headers';
+
+.update-box {
+    align-items: baseline;
+    display: flex;
+    flex-flow: row nowrap;
+    justify-content: center;
+    margin: auto;
+    max-width: $content-max;
+    padding-bottom: 2rem;
+
+    @include width-up-to($phone-max) {
+        align-items: center;
+        display: block;
+        padding-left: 1.5rem;
+        text-align: center;
+    }
+
+    .cta {
+        background-color: os-color(orange);
+        color: text-color(white);
+        margin-left: 1rem;
+        margin-right: 1rem;
+        padding: 1.25rem 3rem;
+
+        @include width-up-to($phone-max) {
+            margin: 0;
+        }
+    }
+
+    .rss-link {
+        background-color: none;
+        color: os-color(orange);
+        margin: 1.5rem;
+        transform: scale(2);
+    }
+}
