+++ src/app/router.js
@@ -13,8 +13,6 @@ const PAGES = [
     'article',
     'blog',
     'blog/*path',
-    'blog-with-search',
-    'blog-with-search/*path',
     'bookstore-suppliers',
     'confirmation',
     'confirmation/*path',
