+++ test/src/pages/blog/blog.test.js
@@ -11,4 +11,16 @@ describe('blog', () => {
             expect(Reflect.ownKeys(p.articles).length).toBe(93);
         })
     );
+    it('displays summary cards', () => {
+        const card = p.el.querySelector('.card');
+
+        expect(card.querySelector('.link-image').href.length).toBeGreaterThan(10);
+    });
+    it('displays the pinned article', () => {
+        const pa = p.el.querySelector('.pinned-article');
+
+        expect(pa).toBeTruthy();
+        expect(pa.querySelector('.article-blurb')).toBeTruthy();
+        expect(pa.querySelector('.no-such-thing')).toBeFalsy();
+    });
 });
