+++ test/src/pages/blog/search-results.test.js
@@ -1,0 +1,9 @@
+import SearchResults from '~/pages/blog/search-results/search-results';
+
+describe('search-results', () => {
+    const p = new SearchResults();
+
+    it('creates', () => {
+        expect(p).toBeTruthy();
+    });
+});
